requirements.txt
pillow==7.0.0
uWSGI==2.0.18
Flask==1.1.2